--- core/src/main/resources/jenkins/model/DownloadSettings/Warning/message_bg.properties
@@ -1,33 +1,0 @@
-# The MIT License
-#
-# Bulgarian translation: Copyright (c) 2016, Alexander Shopov <ash@kambanaria.org>
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-
-# \ 
-# You currently are using browser-based download to retrieve metadata for Jenkins plugins and tools. \ 
-# This has reliability issues and is not considered fully secure. \ 
-# Consider <a href="{0}/configureSecurity">switching to server-based download</a>.
-blurb=\
- \u0412 \u043c\u043e\u043c\u0435\u043d\u0442\u0430 \u043f\u043e\u043b\u0437\u0432\u0430\u0442\u0435 \u0438\u0437\u0442\u0435\u0433\u043b\u044f\u043d\u0435 \u043d\u0430 \u043c\u0435\u0442\u0430\u0434\u0430\u043d\u043d\u0438\u0442\u0435 \u0437\u0430 \u043f\u0440\u0438\u0441\u0442\u0430\u0432\u043a\u0438\u0442\u0435 \u0438 \u0438\u043d\u0441\u0442\u0440\u0443\u043c\u0435\u043d\u0442\u0438\u0442\u0435 \u043d\u0430\
- Jenkins \u043f\u0440\u0435\u0437 \u0431\u0440\u0430\u0443\u0437\u044a\u0440\u0430. \u0422\u043e\u0432\u0430 \u043d\u0435 \u0435 \u0441\u044a\u0432\u0441\u0435\u043c \u043d\u0430\u0434\u0435\u0436\u0434\u043d\u043e, \u0430 \u0438 \u043d\u0435 \u0435 \u0441\u044a\u0432\u0441\u0435\u043c \u0441\u0438\u0433\u0443\u0440\u043d\u043e.\
- \u041f\u043e\u043c\u0438\u0441\u043b\u0435\u0442\u0435 \u0437\u0430 \u0432\u0430\u0440\u0438\u0430\u043d\u0442\u0430 \u043d\u0430 <a href="{0}/configureSecurity">\u0438\u0437\u0442\u0435\u0433\u043b\u044f\u043d\u0438\u044f \u043f\u0440\u0435\u0437\
- \u0441\u044a\u0440\u0432\u044a\u0440\u0430</a>.
-Dismiss=\
- \u041e\u0442\u043a\u0430\u0437
