--- core/src/main/resources/jenkins/model/DownloadSettings/Warning/message_lt.properties
@@ -1,5 +1,0 @@
-blurb=\
-    J\u016bs dabar naudojate nar\u0161ykl\u0117s atsiuntim\u0105, kad gautum\u0117te metaduomenis Jenkinso priedams ir \u012frankiams. \
-    Tai turi patikimumo problem\u0173 ir n\u0117ra visi\u0161kai saugu. \
-    Apsvarstykite <a href="{0}/configureSecurity">persijungim\u0105 \u012f serverio-pus\u0117s atsiuntim\u0105</a>.
-Dismiss=Praleisti
