--- core/src/main/resources/jenkins/model/DownloadSettings/Warning/message_sr.properties
@@ -1,7 +1,0 @@
-# This file is under the MIT License by authors
-
-blurb=\
-    \u0422\u0440\u0435\u043D\u0443\u0442\u043D\u043E \u0432\u0435\u0431-\u043F\u0440\u0435\u0433\u043B\u0435\u0434\u0430\u0447\u0435\u043C \u043F\u0440\u0435\u0443\u0437\u0438\u0430\u0442\u0435 \u043C\u0435\u0442\u0430\u0434\u0430\u0442\u043E\u0442\u0435\u043A\u0435 \u043E Jenkins \u043C\u043E\u0434\u0443\u043B\u0438\u043C\u0430 \u0438 \u0430\u043B\u0430\u0442\u0438\u043C\u0430. \
-    \u0418\u043C\u0430 \u043F\u0440\u043E\u0431\u043B\u0435\u043C\u0430\u0442\u0430 \u0441\u0430 \u043E\u0432\u0438\u043C \u043C\u0435\u0442\u043E\u0434\u043E\u043C, \u0438 \u043D\u0435 \u0441\u043C\u0430\u0442\u0440\u0430 \u0441\u0435 \u043F\u043E\u0442\u043F\u0443\u043D\u043E \u0441\u0438\u0433\u0443\u0440\u043D\u0438\u043C. \
-    \u0420\u0430\u0437\u043C\u043E\u0442\u0440\u0438\u0442\u0435 <a href="{0}/configureSecurity">\u043F\u0440\u0435\u0443\u0437\u0438\u043C\u0430\u045A\u0435 \u0441\u0430 \u0441\u0435\u0440\u0432\u0435\u0440\u043E\u043C</a>.
-Dismiss=\u041E\u0442\u043A\u0430\u0436\u0438
