--- core/src/main/resources/jenkins/model/DownloadSettings/config_sr.properties
@@ -1,4 +1,0 @@
-# This file is under the MIT License by authors
-
-Use\ Browser=\u041A\u043E\u0440\u0438\u0441\u0442\u0438 \u0432\u0435\u0431-\u043F\u0440\u0435\u0433\u043B\u0435\u0434\u0430\u0447
-Use\ browser\ for\ metadata\ download=\u041F\u0440\u0435\u0443\u0437\u0438\u043C\u0430\u0458 \u043C\u0435\u0442\u0430\u0434\u0430\u0442\u043E\u0442\u0435\u043A\u0435 \u0432\u0435\u0431-\u043F\u0440\u0435\u0433\u043B\u0435\u0434\u0430\u0447\u0435\u043C
