--- core/src/main/resources/jenkins/model/DownloadSettings/help-useBrowser.html
@@ -1,11 +1,0 @@
-<p>
-    Check to force the user’s browser to download metadata (lists of available plugins, tools, etc.)
-    rather than Jenkins itself doing it.
-    Actual file downloads (plugins, tools) will still happen from Jenkins itself,
-    but this can be used to at least <em>see</em> new metadata when Jenkins cannot access the Internet
-    (but your own browser can, perhaps using some special proxy that Jenkins is not configured to use).
-</p>
-<p>
-    Use of browser mode is discouraged.
-    It has robustness problems and has been limited to users with Overall/Administer.
-</p>
