--- core/src/main/resources/jenkins/model/DownloadSettings/help-useBrowser_de.html
@@ -1,12 +1,0 @@
-<div>
-    Aktivieren Sie diese Einstellung, wenn Sie Metadaten (Listen verfügbarer Plugins, 
-    Tools etc.) anstelle von Jenkins selbst mit dem Browser herunterladen möchten.
-    
-    Die Downloads der in den Metadaten beschriebenen Dateien werden nach wie vor von 
-    Jenkins übernommen.
-    
-    Diese Einstellung ist nützlich, wenn Jenkins ohne Internetzugriff betrieben wird,
-    Sie aber über Ihren Browser durch besondere Proxy- oder Firewalleinstellungen auf 
-    das Internet zugreifen können.
-    In diesen Fällen können Sie neue Metadaten zumindest <em>sehen</em>.
-</div>
