+++ core/src/main/resources/jenkins/model/Messages.properties
@@ -70,7 +70,6 @@ BlockedBecauseOfBuildInProgress.shortDescription=Build #{0} is already in progre
 BlockedBecauseOfBuildInProgress.ETA=\ (ETA: {0})
 BuildDiscarderProperty.displayName=Discard old builds
 
-DownloadSettings.Warning.DisplayName=Browser-based metadata download
 EnforceSlaveAgentPortAdministrativeMonitor.displayName=Enforce TCP Agent Port
 CLI.disable-job.shortDescription=Disables a job.
 CLI.enable-job.shortDescription=Enables a job.
