+++ core/src/main/resources/jenkins/model/Messages_de.properties
@@ -60,7 +60,6 @@ BlockedBecauseOfBuildInProgress.shortDescription=Build #{0} ist bereits in Arbei
 BlockedBecauseOfBuildInProgress.ETA=\ (ETA: {0})
 BuildDiscarderProperty.displayName=Alte Builds verwerfen
 IdStrategy.CaseSensitive.DisplayName=Gro\u00DF- und Kleinschreibung beachtend
-DownloadSettings.Warning.DisplayName=Browser-basierter Download von Metadaten
 CLI.shutdown.shortDescription=F\u00E4hrt Jenkins sofort herunter.
 EnforceSlaveAgentPortAdministrativeMonitor.displayName=Agent-Port wird bei Neustart zur\u00FCckgesetzt
 CauseOfInterruption.ShortDescription=Abgebrochen von {0}
