+++ core/src/main/resources/jenkins/model/Messages_it.properties
@@ -26,7 +26,6 @@ Mailer.Address.Not.Configured=Indirizzo non ancora configurato <nobody@nowhere>
 IdStrategy.CaseSensitiveEmailAddress.DisplayName=Distinzione maiuscole/minuscole (indirizzo di posta elettronica)
 CauseOfInterruption.ShortDescription=Interrotto da {0}
 CLI.shutdown.shortDescription=Spegne immediatamente il server Jenkins.
-DownloadSettings.Warning.DisplayName=Scaricamento metadati basato sul browser
 IdStrategy.CaseSensitive.DisplayName=Differenza maiuscole/minuscole
 Hudson.NotUsesUTF8ToDecodeURL=\
     Il container non utilizza UTF-8 per decodificare gli URL. Se si utilizzano caratteri non ASCII come nomi \
