+++ test/src/test/java/hudson/model/DownloadServiceTest.java
@@ -1,7 +1,6 @@
 package hudson.model;
 
 import hudson.model.DownloadService.Downloadable;
-import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
@@ -11,12 +10,9 @@ import java.util.TreeSet;
 import hudson.tasks.Maven;
 import hudson.tools.DownloadFromUrlInstaller;
 import hudson.tools.ToolInstallation;
-import jenkins.model.DownloadSettings;
 import net.sf.json.JSONObject;
-import org.jvnet.hudson.test.Issue;
 import org.jvnet.hudson.test.HudsonTestCase;
 import org.jvnet.hudson.test.WithoutJenkins;
-import org.kohsuke.stapler.StaplerResponse;
 
 /**
  * @author Kohsuke Kawaguchi
